chore: move toPath to its own utils file
