+++ client/src/core/router/utils.ts
@@ -1,0 +1,1 @@
+export const toPath = (...paths: string[]) => `/${paths.join("/")}`;
