+++ client/src/feat/auth/logout-callback.tsx
@@ -1,6 +1,6 @@
 import { Navigate } from "@solidjs/router";
 import { CoreRoute } from "core/constants/core-route.enum";
-import { toPath } from "core/router/route";
+import { toPath } from "core/router/utils";
 import type { Component, ParentProps } from "solid-js";
 
 export const LogoutCallback: Component<ParentProps> = () => (
